[Extension] Display log output in Figma Console pad window
When running an action in Visual Studio for Mac a Figma Console
pad window is created, which can be opened from the status bar
itself, or directly. This pad window displays any messages
sent to the LoggingService. If there is no active action the
LoggingService calls will result in messages stored in
Visual Studio for Mac's IDE log.