[Controls.Cocoa] Refactor node skipping
Instead of checking all parent nodes to see if the children are
skipped only the main node is checked for the !skip attribute.
Instead the CodeRenderService now has a new SkipChildNodes method
which is overriden in the NativeViewCodeService which checks to
see if the node has the !skip attribute and then does not generate
code for any child nodes.